feat: add dynamic metadata per tour page (LES-79)
