+++ src/app/tours/[name]/tour-client.tsx
@@ -1,0 +1,242 @@
+'use client'
+
+import { Button } from '@/components/ui/button'
+import { Card, CardContent, CardHeader } from '@/components/ui/card'
+import { TourMediaCarousel } from '@/components/ui/tour-media-carousel'
+import { CONTACT } from '@/lib/data'
+import { images } from '@/lib/images'
+import { ArrowLeft, Check, Clock, MapPin } from 'lucide-react'
+import Image from 'next/image'
+import Link from 'next/link'
+import { useEffect, useState } from 'react'
+
+interface Activity {
+  description: string
+  duration?: string
+  image: string
+  includes?: string[]
+  price?: number
+  title: string
+}
+
+export interface Tour {
+  activities: Activity[]
+  description: string
+  duration: string
+  highlights: string[]
+  href: string
+  image: string
+  images?: Array<{
+    alt: string
+    type: 'image' | 'video'
+    url: string
+  }>
+  place: string
+  price: number
+  title: string
+}
+
+export default function TourClient({ tour }: { tour: Tour }) {
+  const [selectedActivities, setSelectedActivities] = useState<string[]>([])
+  const [message, setMessage] = useState<string>('')
+
+  const getSelectedActivityPrice = (activityTitle: string) => {
+    const activity = tour.activities?.find(
+      (activity) => activity.title === activityTitle
+    )
+    return Number(activity?.price ?? 0)
+  }
+
+  const totalPrice =
+    tour.price +
+    selectedActivities.reduce(
+      (acc, title) => acc + getSelectedActivityPrice(title),
+      0
+    )
+
+  useEffect(() => {
+    setMessage(
+      `Hello, I would like to get more information about the tour ${tour.title}${selectedActivities.length > 0 ? `, with the activities: ${selectedActivities.join(', ')}` : ''}. With an approximate value of ${totalPrice}`
+    )
+  }, [tour, selectedActivities, totalPrice])
+
+  const handleActivityToggle = (activityTitle: string) => {
+    setSelectedActivities((prev) => {
+      if (prev.includes(activityTitle)) {
+        return prev.filter((title) => title !== activityTitle)
+      }
+      return [...prev, activityTitle]
+    })
+  }
+
+  return (
+    <div className="mx-auto my-14 flex flex-col items-center justify-center gap-12 md:w-6/12">
+      <Link
+        href="/tours"
+        className="text-muted-foreground hover:text-foreground flex items-center gap-2 self-start text-sm transition-colors duration-200"
+      >
+        <ArrowLeft className="h-4 w-4" />
+        Back to tours
+      </Link>
+      <Card className="mx-auto flex w-11/12 flex-col pt-0 md:w-full">
+        <CardHeader className="px-0 pt-0">
+          {tour.images ? (
+            <TourMediaCarousel items={tour.images} />
+          ) : (
+            <Image
+              src={tour.image}
+              alt={tour.place}
+              width={500}
+              height={500}
+              className="aspect-[4/3] w-full rounded-t-[14px] object-cover"
+            />
+          )}
+        </CardHeader>
+        <CardContent className="flex flex-col gap-4">
+          <div className="flex flex-wrap items-center gap-4">
+            <div className="text-muted-foreground flex items-center gap-1">
+              <MapPin />
+              <p>{tour.place}</p>
+            </div>
+            <div className="text-muted-foreground flex items-center gap-1">
+              <Clock />
+              <p>{tour.duration}</p>
+            </div>
+          </div>
+          <h2 className="text-2xl font-bold">{tour.title}</h2>
+          <p className="text-muted-foreground">{tour.description}</p>
+          <h3 className="text-xl font-bold">Highlights:</h3>
+          <ul className="grid list-inside list-disc gap-2 md:grid-cols-2">
+            {tour.highlights.map((highlight) => (
+              <li key={highlight}>{highlight}</li>
+            ))}
+          </ul>
+        </CardContent>
+      </Card>
+
+      {tour.activities.length > 0 && (
+        <h2 className="ml-4 self-start text-2xl font-bold">
+          Available activities:
+        </h2>
+      )}
+      <div className="flex flex-col gap-4">
+        {tour.activities.length > 0
+          ? tour.activities.map((activity) => (
+              <Card
+                key={activity.title}
+                className="mx-auto flex w-11/12 flex-col py-0 md:w-full md:flex-row"
+              >
+                <Image
+                  src={activity.image}
+                  alt={activity.title}
+                  width={200}
+                  height={200}
+                  className="w-full rounded-t-[14px] object-cover md:w-4/12 md:rounded-l-[14px] md:rounded-r-none"
+                />
+                <div className="flex flex-col gap-2 p-4">
+                  <div className="flex w-full items-center justify-between gap-4">
+                    <h3 className="text-xl font-bold">{activity.title}</h3>
+                    <p className="text-muted-foreground">{activity.duration}</p>
+                  </div>
+                  <p className="text-muted-foreground">
+                    {activity.description}
+                  </p>
+                  {activity.includes && (
+                    <div>
+                      <p className="font-bold">Includes:</p>
+                      <ul className="grid grid-cols-2 gap-2">
+                        {activity.includes.map((include: string) => (
+                          <li key={include} className="flex items-center gap-2">
+                            <Check className="text-primary h-4 w-4" />
+                            {include}
+                          </li>
+                        ))}
+                      </ul>
+                    </div>
+                  )}
+                  <div className="flex items-center justify-between">
+                    <p className="text-secondary text-xl font-bold">
+                      ${Number(activity.price).toLocaleString()}
+                    </p>
+                    <Button
+                      variant={
+                        selectedActivities.includes(activity.title)
+                          ? 'outline'
+                          : 'default'
+                      }
+                      onClick={() => handleActivityToggle(activity.title)}
+                    >
+                      {selectedActivities.includes(activity.title)
+                        ? 'Remove from my experience'
+                        : 'Add to my experience'}
+                    </Button>
+                  </div>
+                </div>
+              </Card>
+            ))
+          : null}
+      </div>
+      <Card className="mx-auto w-11/12 px-4 md:w-full">
+        <CardHeader>
+          <h2 className="text-2xl font-bold">Summary of your experience</h2>
+          <div className="flex flex-col gap-4">
+            <div className="flex items-center justify-between">
+              <p className="text-muted-foreground">Base price:</p>
+              <p className="text-lg font-bold">
+                ${Number(tour.price).toLocaleString()} / person
+              </p>
+            </div>
+            {selectedActivities.length > 0 && (
+              <>
+                <div className="border-t pt-4">
+                  <p className="mb-2 font-bold">Selected activities:</p>
+                  {selectedActivities.map((title) => (
+                    <div
+                      key={title}
+                      className="flex items-center justify-between"
+                    >
+                      <p className="text-muted-foreground">{title}</p>
+                      <p className="font-bold">
+                        $
+                        {Number(
+                          getSelectedActivityPrice(title)
+                        ).toLocaleString()}
+                      </p>
+                    </div>
+                  ))}
+                </div>
+                <div className="border-t pt-4">
+                  <div className="flex items-center justify-between">
+                    <p className="text-lg font-bold">Total:</p>
+                    <p className="text-lg font-bold">
+                      ${Number(totalPrice).toLocaleString()}
+                    </p>
+                  </div>
+                </div>
+              </>
+            )}
+          </div>
+        </CardHeader>
+        <Button
+          className="flex w-full items-center gap-2 bg-green-500 text-white hover:bg-green-600"
+          asChild
+        >
+          <Link
+            href={`https://wa.me/${CONTACT.phone}?text=${encodeURIComponent(message)}`}
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <Image
+              src={images.whatsapp}
+              alt="whatsapp"
+              width={20}
+              height={20}
+            />
+            <p>Consult by Whatsapp</p>
+          </Link>
+        </Button>
+      </Card>
+      <div className="flex justify-center"></div>
+    </div>
+  )
+}
